[fix] 방에 참여한 경우와 아닌 경우에 따라 로직 수정

+++ src/main/java/hous/server/service/user/UserService.java
@@ -12,6 +12,7 @@ import hous.server.domain.personality.PersonalityColor;
 import hous.server.domain.personality.repository.PersonalityRepository;
 import hous.server.domain.room.Participate;
 import hous.server.domain.room.Room;
+import hous.server.domain.room.repository.ParticipateRepository;
 import hous.server.domain.room.repository.RoomRepository;
 import hous.server.domain.todo.Todo;
 import hous.server.domain.todo.repository.DoneRepository;
@@ -59,6 +60,7 @@ public class UserService {
     private final DoneRepository doneRepository;
     private final TodoRepository todoRepository;
     private final RoomRepository roomRepository;
+    private final ParticipateRepository participateRepository;
 
     private final BadgeService badgeService;
 
@@ -143,17 +145,20 @@ public class UserService {
 
     public void deleteUser(Long userId) {
         User user = UserServiceUtils.findUserById(userRepository, userId);
-        Room room = RoomServiceUtils.findParticipatingRoom(user);
         Onboarding me = user.getOnboarding();
+        Room room = me.getParticipates().get(0).getRoom();
 
-        List<Todo> todos = room.getTodos();
-        List<Todo> myTodos = TodoServiceUtils.filterAllDaysUserTodos(todos, me);
-        RoomServiceUtils.deleteMyTodosTakeMe(takeRepository, doneRepository, todoRepository, myTodos, me, room);
+        if (room != null) {
+            List<Todo> todos = room.getTodos();
+            List<Todo> myTodos = TodoServiceUtils.filterAllDaysUserTodos(todos, me);
+            RoomServiceUtils.deleteMyTodosTakeMe(takeRepository, doneRepository, todoRepository, myTodos, me, room);
+            RoomServiceUtils.deleteParticipateUser(participateRepository, roomRepository, me, room);
+            if (room.getParticipates().isEmpty()) {
+                roomRepository.delete(room);
+            }
+        }
 
         userRepository.delete(user);
-        if (room.getParticipates().isEmpty()) {
-            roomRepository.delete(room);
-        }
     }
 
     public void acquireFeedbackBadge(Long userId) {
